+++ lib/bitstamp_client/factories.rb
@@ -1,19 +1,19 @@
-FactoryGirl.define do
+FactoryBot.define do
 
   factory :bitstamp_client_order_book, class: "BitstampClient::OrderBook" do
     timestamp { Time.now.to_i }
 
     bids do
       [
-        FactoryGirl.build(:bitstamp_client_order),
-        FactoryGirl.build(:bitstamp_client_order),
+        FactoryBot.build(:bitstamp_client_order),
+        FactoryBot.build(:bitstamp_client_order),
       ]
     end
 
     asks do
       [
-        FactoryGirl.build(:bitstamp_client_order),
-        FactoryGirl.build(:bitstamp_client_order),
+        FactoryBot.build(:bitstamp_client_order),
+        FactoryBot.build(:bitstamp_client_order),
       ]
     end
   end